backend/requirements.txt
# =========================
# Core FastAPI Stack
# =========================
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
python-dotenv>=1.0.1

# =========================
# Data Validation
# =========================
pydantic>=2.6.0

# ⚠️ Pandas works, but avoid heavy ops
pandas>=2.2.0
numpy>=1.26.0

# =========================
# Database (MongoDB)
# =========================
motor>=3.4.0
pymongo>=4.7.0
dnspython>=2.6.0

# =========================
# Authentication & Security
# =========================
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
bcrypt>=4.1.2

# =========================
# FastAPI Internals
# =========================
starlette>=0.36.0

# =========================
# Translation (Python 3.13 SAFE)
# =========================
deep-translator>=1.11.4
translate>=3.6.1
translators>=5.9.0
langdetect>=1.0.9

# =========================
# HTTP Clients
# =========================
requests>=2.32.0
aiohttp>=3.9.5
httpx>=0.27.0

# =========================
# Caching
# =========================
cachetools>=5.5.0
redis>=5.0.4

# =========================
# AI APIs (Cloud-based, safe)
# =========================
google-generativeai>=0.5.0

# =========================
# Image Processing
# =========================
Pillow>=10.3.0

# =========================
# Speech Recognition (NO PyAudio)
# =========================
SpeechRecognition>=3.10.1

# =========================
# Testing
# =========================
pytest>=8.0.0
pytest-asyncio>=0.23.5

# =========================
# Logging
# =========================
structlog>=24.1.0
PyJWT
pyotp
qrcode
webauthn
tensorflow
groq
